refactor: use application layout components from react-components
