--- demo/src/components/ApplicationLayout/AppAside/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppAside";
