--- demo/src/components/ApplicationLayout/AppNavigation/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppNavigation";
