--- demo/src/components/ApplicationLayout/AppNavigationBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppNavigationBar";
