--- demo/src/components/ApplicationLayout/AppStatus/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppStatus";
