--- demo/src/components/ApplicationLayout/Application/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Application";
