--- demo/src/components/ApplicationLayout/index.ts
@@ -1,7 +1,0 @@
-export { default } from "./ApplicationLayout";
-export { default as AppAside } from "./AppAside";
-export { default as Application } from "./Application";
-export { default as AppMain } from "./AppMain";
-export { default as AppNavigation } from "./AppNavigation";
-export { default as AppNavigationBar } from "./AppNavigationBar";
-export { default as AppStatus } from "./AppStatus";
