--- demo/src/components/SideNavigation/SideNavigationItem/index.ts
@@ -1,4 +1,0 @@
-export {
-  default,
-  type Props as SideNavigationItemProps,
-} from "./SideNavigationItem";
