--- demo/src/components/SideNavigation/SideNavigationLink/index.ts
@@ -1,5 +1,0 @@
-export {
-  default,
-  type LinkDefaultElement as SideNavigationLinkDefaultElement,
-  type Props as SideNavigationLinkProps,
-} from "./SideNavigationLink";
