--- demo/src/components/SideNavigation/index.ts
@@ -1,7 +1,0 @@
-export { default } from "./SideNavigation";
-export type { Props as SideNavigationProps } from "./SideNavigation";
-export { default as SideNavigationItem } from "./SideNavigationItem";
-export {
-  default as SideNavigationLink,
-  type SideNavigationLinkDefaultElement,
-} from "./SideNavigationLink";
