+++ demo/src/main.tsx
@@ -1,8 +1,8 @@
+import { Panel } from "@canonical/react-components";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { Link, RouterProvider, createBrowserRouter } from "react-router-dom";
 
-import Panel from "components/Panel";
 import ReBACAdmin from "components/ReBACAdmin";
 import { logger } from "utils";
