+++ package.json
@@ -66,7 +66,7 @@
     "react-hot-toast": "2.4.1"
   },
   "devDependencies": {
-    "@canonical/react-components": "0.56.0",
+    "@canonical/react-components": "0.59.0",
     "@faker-js/faker": "8.4.1",
     "@mswjs/http-middleware": "0.10.1",
     "@tanstack/eslint-plugin-query": "5.32.1",
