+++ src/components/FormPanel/types.ts
@@ -1,6 +1,6 @@
+import type { PanelProps } from "@canonical/react-components";
 import type { FormikValues } from "formik";
 
-import type { PanelProps } from "components/Panel";
 import type { Props as PanelFormProps } from "components/PanelForm";
 
 export type Props<F extends FormikValues> = {
