--- src/components/Panel/index.ts
@@ -1,5 +1,0 @@
-export {
-  default,
-  type Props as PanelProps,
-  type LogoDefaultElement as PanelLogoDefaultElement,
-} from "./Panel";
